# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "此 API 文件提供社區通知系統的後端介面說明，包含使用者管理、設施預約、社區申請及訊息通知等功能。使用 JWT Bearer 進行認證。",
        "title": "社區通知系統 API 文件 (Community Notification System)",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/platform/getlist": {
            "get": {
                "description": "取得所有可用平台名稱，供前端顯示於選項列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "取得平台列表",
                "responses": {
                    "200": {
                        "description": "成功取得平台清單",
                        "schema": {
                            "$ref": "#/definitions/platform.PlatformListResponse"
                        }
                    },
                    "500": {
                        "description": "取得平台資料失敗",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Response500Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "error": {
                    "type": "string",
                    "example": "Internal server error"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1000
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
        "platform.PlatformListResponse": {
            "type": "object",
            "properties": {
                "platforms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/platform.PlatformSummary"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "platform.PlatformSummary": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}